Redesigned mobile view of projects
