+++ src/components/energies/energies.component.jsx
@@ -4,14 +4,11 @@ import { useUpdateEffect } from "react-use";
 // UI Core
 import { makeStyles } from "@material-ui/core/styles";
 // Selectors
-import { selectSessionsComplete } from "../../redux/projects/projects.selectors";
 import { selectEnergies } from "../../redux/energies/energies.selectors";
-// Actions
-import { setEnergies } from "../../redux/energies/energies.actions.js";
 // Components
 import ProgressBar from "../progress-bar/progress-bar.component.jsx";
 
-const Energies = ({ energies, setEnergies, sessionsComplete }) => {
+const Energies = ({ energies }) => {
     const { body, emotions, mind, soul } = energies;
 
     const useStyles = makeStyles({
@@ -27,14 +24,7 @@ const Energies = ({ energies, setEnergies, sessionsComplete }) => {
     const classes = useStyles();
 
     // TODO: Connect to API
-    useUpdateEffect(() => {
-        setEnergies({
-            body: body - 5,
-            emotions: emotions - 10,
-            mind: mind - 15,
-            soul: soul - 5,
-        });
-    }, [sessionsComplete]);
+    useUpdateEffect(() => {}, [body, emotions, mind, soul]);
 
     return (
         <div className={classes.energies}>
@@ -48,11 +38,6 @@ const Energies = ({ energies, setEnergies, sessionsComplete }) => {
 
 const mapStateToProps = state => ({
     energies: selectEnergies(state),
-    sessionsComplete: selectSessionsComplete(state),
 });
 
-const mapDispatchToProps = dispatch => ({
-    setEnergies: value => dispatch(setEnergies(value)),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(Energies);
+export default connect(mapStateToProps)(Energies);
