+++ src/components/project-timer/project-timer.component.jsx
@@ -13,6 +13,8 @@ import {
     selectSessionsComplete,
     selectBreakInProgress,
 } from "../../redux/projects/projects.selectors";
+import { selectEnergies } from "../../redux/energies/energies.selectors";
+import { selectStats } from "../../redux/stats/stats.selectors";
 // Components
 import TimerDisplays from "../timer-displays/timer-displays.component.jsx";
 // Actions
@@ -21,6 +23,8 @@ import {
     setSessionInProgress,
     setSessionsComplete,
 } from "../../redux/projects/projects.actions";
+import { setEnergies } from "../../redux/energies/energies.actions";
+import { setStats } from "../../redux/stats/stats.actions";
 
 const ProjectTimer = ({
     method,
@@ -33,7 +37,14 @@ const ProjectTimer = ({
     minuteLeftSound,
     breakCompleteSound,
     sessionEndSound,
+    setEnergies,
+    energies,
+    setStats,
+    stats,
 }) => {
+    const { body, emotions, mind, soul } = energies;
+    const { strength, creativity, intelligence, fluency } = stats;
+
     const [sessionTime, setSessionTime] = useState(duration(method, "minutes"));
     const [localSession, setLocalSession] = useState(false);
 
@@ -58,7 +69,7 @@ const ProjectTimer = ({
             sessionEndSound.play();
         }
     }, [sessionTime, sessionEndSound]);
-    // Session Time
+    // Session Management
     useEffect(() => {
         setSessionTime(duration(method, "minutes"));
     }, [method]);
@@ -96,6 +107,22 @@ const ProjectTimer = ({
                           } else {
                               // Stop timer
                               setLocalSession(false);
+
+                              //Update Energies
+                              setEnergies({
+                                  body: body - 5,
+                                  emotions: emotions - 10,
+                                  mind: mind - 15,
+                                  soul: soul - 5,
+                              });
+
+                              // Update Stats
+                              setStats({
+                                  strength: strength + 0,
+                                  creativity: creativity + 5,
+                                  intelligence: intelligence + 10,
+                                  fluency: fluency + 0,
+                              });
 
                               // Update sessions count
                               const newCount = sessionsComplete + 1;
@@ -135,6 +162,16 @@ const ProjectTimer = ({
         setSessionsComplete,
         sessionEndSound,
         minuteLeftSound,
+        setEnergies,
+        setStats,
+        strength,
+        creativity,
+        intelligence,
+        fluency,
+        body,
+        mind,
+        emotions,
+        soul,
     ]);
 
     useEffect(() => {
@@ -200,12 +237,16 @@ const mapStateToProps = state => ({
     sessionEndSound: state.uifx.projectSounds.sessionEndSound,
     minuteLeftSound: state.uifx.projectSounds.minuteLeftSound,
     breakCompleteSound: state.uifx.projectSounds.breakCompleteSound,
+    energies: selectEnergies(state),
+    stats: selectStats(state),
 });
 
 const mapDispatchToProps = dispatch => ({
     setBreakInProgress: value => dispatch(setBreakInProgress(value)),
     setSessionInProgress: value => dispatch(setSessionInProgress(value)),
     setSessionsComplete: value => dispatch(setSessionsComplete(value)),
+    setEnergies: value => dispatch(setEnergies(value)),
+    setStats: value => dispatch(setStats(value)),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(ProjectTimer);
