+++ src/redux/stats/stats.selectors.js
@@ -1,0 +1,1 @@
+export const selectStats = state => state.stats;
